solution for Two/Eleven
This one's a bit messy.  It prints two things when run; the first is
which of ECB/CBC was run inside the oracle, and the second is which the
oracle detector thinks was run.  I haven't seen it fail, but thus is
randomness - there may well be a set against which it fails (I suspect
it may misdiagnose CBC as ECB under exceptional circumstances).